requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Trading Dependencies
requests==2.31.0
pyotp==2.9.0
python-dotenv==1.0.0
asyncio==3.4.3

# Data & Utilities
pydantic==2.5.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Development
watchfiles==0.21.0
httpx==0.25.1